services/storage-writer/requirements.txt
fastapi
uvicorn[standard]
kafka-python
psycopg2-binary
prometheus-client
python-dotenv